Refactored verify account to a better logic
